Move chain-generation into separate crate + add more tests
