+++ ergo-chain-generation/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "ergo-chain-generation"
+version = "0.14.0"
+license = "CC0-1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base16 = "0.2.1"
+blake2 = "0.9"
+byteorder = "1"
+ergotree-ir = { version = "^0.14.0", path = "../ergotree-ir", features = ["arbitrary"] }
+ergotree-interpreter = { version = "^0.14.0", path = "../ergotree-interpreter", features = ["arbitrary"] }
+ergo-lib = {version = "^0.14.0", path = "../ergo-lib"}
+num-bigint = "0.4.0"
+rand = "0.8.4"
+sigma-test-util = { version = "^0.3.0", path = "../sigma-test-util" }
+ergo-nipopow = { version = "^0.1", path = "../ergo-nipopow" }
